Add logical layer to crypto that represents spend zkp leaf nodes in a more general way.
