+++ crypto/src/main/java/org/cryptimeleon/incentive/crypto/IncentiveSystem.java
@@ -15,9 +15,9 @@ import org.cryptimeleon.incentive.crypto.model.keys.user.UserPublicKey;
 import org.cryptimeleon.incentive.crypto.model.keys.user.UserSecretKey;
 import org.cryptimeleon.incentive.crypto.model.messages.JoinRequest;
 import org.cryptimeleon.incentive.crypto.model.messages.JoinResponse;
-import org.cryptimeleon.incentive.crypto.proof.spend.SpendDeductZkp;
-import org.cryptimeleon.incentive.crypto.proof.spend.SpendDeductZkpCommonInput;
-import org.cryptimeleon.incentive.crypto.proof.spend.SpendDeductZkpWitnessInput;
+import org.cryptimeleon.incentive.crypto.proof.spend.zkp.SpendDeductZkp;
+import org.cryptimeleon.incentive.crypto.proof.spend.zkp.SpendDeductZkpCommonInput;
+import org.cryptimeleon.incentive.crypto.proof.spend.zkp.SpendDeductZkpWitnessInput;
 import org.cryptimeleon.incentive.crypto.proof.wellformedness.CommitmentWellformednessCommonInput;
 import org.cryptimeleon.incentive.crypto.proof.wellformedness.CommitmentWellformednessProtocol;
 import org.cryptimeleon.incentive.crypto.proof.wellformedness.CommitmentWellformednessWitness;
