+++ crypto/src/main/java/org/cryptimeleon/incentive/crypto/Setup.java
@@ -38,23 +38,18 @@ import java.util.stream.LongStream;
 public class Setup {
     // Which AES version to use for the PRF-to-ZN
     public static final int HASH_THEN_PRF_AES_KEY_LENGTH = 256;
-
+    // The size of the metadata base vector of a token
+    public static final int H_SIZE_WITHOUT_POINTS = 6;
     // (1/2)^OVERSUBSCRIPTION is a lower bound for probability of the PRF failing
     private static final int HASH_THEN_PRF_OVERSUBSCRIPTION = 128;
-
     // Base for esk decomposition. Requires as many signed digits in the provider's public key, but reduces digits and
     // hence verifications required in the Spend-Deduct protocol for the encryption secret key
     private static final long ESK_DEC_BASE = 256;
-
     // ESK_DEC_BASE^MAX_POINTS_BASE_POWER determines the maximum number of points a user can have.
     // This is due to the CCS range proof used for v>=k in Spend-Deduct
     private static final int MAX_POINTS_BASE_POWER = 6;
-
     // Determines the maximum size of a promotion's points vector under a provider public key
     private static final int MAX_POINTS_VECTOR_SIZE = 10;
-
-    // The size of the metadata base vector of a token
-    public static final int H_SIZE_WITHOUT_POINTS = 6;
 
     /**
      * Generates public parameters from security parameter
