+++ crypto/src/main/java/org/cryptimeleon/incentive/crypto/proof/spend/leaf/MetadataLeaf.java
@@ -1,0 +1,11 @@
+package org.cryptimeleon.incentive.crypto.proof.spend.leaf;
+
+import org.cryptimeleon.incentive.crypto.proof.spend.tree.SpendDeductLeafNode;
+
+public class MetadataLeaf extends SpendDeductLeafNode {
+
+    public MetadataLeaf() {
+        super("MetadataZkp");
+        setHasWitness(true);
+    }
+}
