+++ crypto/src/main/java/org/cryptimeleon/incentive/crypto/proof/spend/zkp/TokenUpdateZkp.java
@@ -1,9 +1,10 @@
-package org.cryptimeleon.incentive.crypto.proof.spend;
+package org.cryptimeleon.incentive.crypto.proof.spend.zkp;
 
 import org.cryptimeleon.craco.protocols.CommonInput;
 import org.cryptimeleon.craco.protocols.SecretInput;
 import org.cryptimeleon.craco.protocols.arguments.sigma.ZnChallengeSpace;
 import org.cryptimeleon.craco.protocols.arguments.sigma.schnorr.DelegateProtocol;
+import org.cryptimeleon.craco.protocols.arguments.sigma.schnorr.LinearExponentStatementFragment;
 import org.cryptimeleon.craco.protocols.arguments.sigma.schnorr.LinearStatementFragment;
 import org.cryptimeleon.craco.protocols.arguments.sigma.schnorr.SendThenDelegateFragment;
 import org.cryptimeleon.craco.protocols.arguments.sigma.schnorr.setmembership.SmallerThanPowerFragment;
@@ -13,31 +14,34 @@ import org.cryptimeleon.incentive.crypto.model.keys.provider.ProviderPublicKey;
 import org.cryptimeleon.math.expressions.exponent.ExponentConstantExpr;
 import org.cryptimeleon.math.structures.cartesian.ExponentExpressionVector;
 import org.cryptimeleon.math.structures.cartesian.GroupElementExpressionVector;
+import org.cryptimeleon.math.structures.cartesian.Vector;
 import org.cryptimeleon.math.structures.groups.GroupElement;
 import org.cryptimeleon.math.structures.rings.zn.Zn;
 
 import java.math.BigInteger;
 
-public class TokenPointsRangeProof extends DelegateProtocol {
+public class TokenUpdateZkp extends DelegateProtocol {
 
-    public TokenPointsRangeProof(IncentivePublicParameters pp, BigInteger[] lowerLimits, BigInteger[] upperLimits, ProviderPublicKey providerPublicKey, PromotionParameters promotionParameters) {
+    // Public parameters
+    final IncentivePublicParameters pp;
+    // Proof that lowerLimits[i] <= newPoints[i] <= upperLimits[i] for all i
+    final BigInteger[] lowerLimits; // null means no limit
+    final BigInteger[] upperLimits; // null means no limit
+    // Proof linear relation between new points and old points
+    // newPoints[i] = a[i]* oldPoints[i] + bVector[i]
+    final BigInteger[] aVector; // null means no statement
+    final BigInteger[] bVector; // null means no statement
+    final ProviderPublicKey providerPublicKey;
+    final PromotionParameters promotionParameters;
+    public TokenUpdateZkp(IncentivePublicParameters pp, BigInteger[] lowerLimits, BigInteger[] upperLimits, BigInteger[] aVector, BigInteger[] bVector, ProviderPublicKey providerPublicKey, PromotionParameters promotionParameters) {
         this.pp = pp;
         this.lowerLimits = lowerLimits;
         this.upperLimits = upperLimits;
+        this.aVector = aVector;
+        this.bVector = bVector;
         this.providerPublicKey = providerPublicKey;
         this.promotionParameters = promotionParameters;
     }
-
-    // Public parameters
-    IncentivePublicParameters pp;
-
-    // Proof that lowerLimits[i] <= points[i] <= upperLimits[i] for all i
-    BigInteger[] lowerLimits; // null means no limit
-    BigInteger[] upperLimits; // null means no limit
-
-    ProviderPublicKey providerPublicKey;
-    PromotionParameters promotionParameters;
-
 
     @Override
     protected SendThenDelegateFragment.SubprotocolSpec provideSubprotocolSpec(CommonInput pCommonInput, SendThenDelegateFragment.SubprotocolSpecBuilder builder) {
@@ -53,6 +57,13 @@ public class TokenPointsRangeProof extends DelegateProtocol {
         var zVar = builder.addZnVariable("z", zn);
         var tVar = builder.addZnVariable("t", zn);
         var pointsVector = ExponentExpressionVector.generate(i -> builder.addZnVariable("points_" + i, zn), promotionParameters.getPointsVectorSize());
+        var newPointsVector = ExponentExpressionVector.generate(i -> builder.addZnVariable("newPoints_" + i, zn), promotionParameters.getPointsVectorSize());
+        var dsrndStar0Var = builder.addZnVariable("dsrndStar0", zn);
+        var dsrndStar1Var = builder.addZnVariable("dsrndStar1", zn);
+        var zStarVar = builder.addZnVariable("zStar", zn);
+        var tStarVar = builder.addZnVariable("tStar", zn);
+        var uStarInverseVar = builder.addZnVariable("uStarInverse", zn);
+        var eskDecVarVector = ExponentExpressionVector.generate(i -> builder.addZnVariable("eskStarUserDec_" + i, zn), pp.getNumEskDigits());
 
 
         // C=(H.pow(t, usk, esk, dsrnd_0, dsrnd_1, points, z), g_1)
@@ -61,13 +72,25 @@ public class TokenPointsRangeProof extends DelegateProtocol {
         ).isEqualTo(commonInput.commitmentC0);
         builder.addSubprotocol("C0", new LinearStatementFragment(commitmentC0Statement));
 
+        // C_1=H.pow(t^*, usk, \sum_i=0^k[esk^*_(usr,i) * base^i], dsrnd^*_0, dsrnd^*_1, z^*, newPointsVector)
+        var powersOfEskDecBase = ExponentExpressionVector.generate(i -> pp.getEskDecBase().pow(BigInteger.valueOf(i)).asExponentExpression(), pp.getNumEskDigits()); // construct vector (eskBase^0, eskBase^1, ...)
+        var exponents = new Vector<>(
+                tStarVar,
+                uskVar,
+                eskDecVarVector.innerProduct(powersOfEskDecBase),
+                dsrndStar0Var,
+                dsrndStar1Var,
+                zStarVar
+        ).concatenate(newPointsVector);
+        builder.addSubprotocol("C0Pre", new LinearStatementFragment(H.innerProduct(exponents).isEqualTo(commonInput.c0Pre.pow(uStarInverseVar))));
+
         // Add range proofs
         for (int i = 0; i < promotionParameters.getPointsVectorSize(); i++) {
             if (lowerLimits[i] != null) {
                 builder.addSubprotocol(
-                        "pointsVector[" + i + "]>=lowerLimits[" + i + "]",
+                        "newPointsVector[" + i + "]>=lowerLimits[" + i + "]",
                         new SmallerThanPowerFragment(
-                                pointsVector.get(i).sub(lowerLimits[i].intValue()),
+                                newPointsVector.get(i).sub(lowerLimits[i].intValue()),
                                 pp.getEskDecBase().asInteger().intValue(),
                                 pp.getMaxPointBasePower(),
                                 pp.getEskBaseSetMembershipPublicParameters()
@@ -76,14 +99,27 @@ public class TokenPointsRangeProof extends DelegateProtocol {
             }
             if (upperLimits[i] != null) {
                 builder.addSubprotocol(
-                        "pointsVector[" + i + "]<=upperLimits[" + i,
+                        "newPointsVector[" + i + "]<=upperLimits[" + i,
                         new SmallerThanPowerFragment(
-                                (new ExponentConstantExpr(upperLimits[i])).sub(pointsVector.get(i)),
+                                (new ExponentConstantExpr(upperLimits[i])).sub(newPointsVector.get(i)),
                                 pp.getEskDecBase().asInteger().intValue(),
                                 pp.getMaxPointBasePower(),
                                 pp.getEskBaseSetMembershipPublicParameters()
                         )
                 );
+            }
+        }
+
+        // Add affine linear statements
+        for (int i = 0; i < promotionParameters.getPointsVectorSize(); i++) {
+            if (aVector[i] != null && bVector[i] != null) {
+                builder.addSubprotocol(
+                        "newPointsVector[" + i + "]=aVector[" + i + "]*pointsVector[" + i + "]+bVector[" + i + "]",
+                        new LinearExponentStatementFragment(
+                                pointsVector.get(i).mul(aVector[i]).add(bVector[i]).isEqualTo(newPointsVector.get(i)), zn
+                        )
+                );
+
             }
         }
         return builder.build();
@@ -98,13 +134,22 @@ public class TokenPointsRangeProof extends DelegateProtocol {
         builder.putWitnessValue("usk", secretInput.usk);
         builder.putWitnessValue("dsrnd0", secretInput.dsrnd0);
         builder.putWitnessValue("dsrnd1", secretInput.dsrnd1);
+        builder.putWitnessValue("dsrndStar0", secretInput.dsrndStar0);
+        builder.putWitnessValue("dsrndStar1", secretInput.dsrndStar1);
         builder.putWitnessValue("z", secretInput.z);
         builder.putWitnessValue("t", secretInput.t);
+        builder.putWitnessValue("zStar", secretInput.zStar);
+        builder.putWitnessValue("tStar", secretInput.tStar);
+        builder.putWitnessValue("uStarInverse", secretInput.uStar.inv());
+
+        for (int i = 0; i < pp.getNumEskDigits(); i++) {
+            builder.putWitnessValue("eskStarUserDec_" + i, (Zn.ZnElement) secretInput.eskStarUserDec.get(i));
+        }
 
         for (int i = 0; i < promotionParameters.getPointsVectorSize(); i++) {
             builder.putWitnessValue("points_" + i, (Zn.ZnElement) secretInput.pointsVector.get(i));
+            builder.putWitnessValue("newPoints_" + i, (Zn.ZnElement) secretInput.newPointsVector.get(i));
         }
-
         return builder.build();
     }
