+++ crypto/src/test/java/org/cryptimeleon/incentive/crypto/ProtocolIntegrationTest.java
@@ -7,8 +7,8 @@ import org.cryptimeleon.incentive.crypto.model.SpendRequest;
 import org.cryptimeleon.incentive.crypto.model.SpendResponse;
 import org.cryptimeleon.incentive.crypto.model.messages.JoinRequest;
 import org.cryptimeleon.incentive.crypto.model.messages.JoinResponse;
-import org.cryptimeleon.incentive.crypto.proof.spend.MetadataZkp;
 import org.cryptimeleon.incentive.crypto.proof.spend.SpendHelper;
+import org.cryptimeleon.incentive.crypto.proof.spend.zkp.MetadataZkp;
 import org.cryptimeleon.incentive.crypto.proof.wellformedness.CommitmentWellformednessProtocol;
 import org.cryptimeleon.math.structures.cartesian.Vector;
 import org.cryptimeleon.math.structures.rings.RingElement;
