+++ crypto/src/test/java/org/cryptimeleon/incentive/crypto/SpendDeductTest.java
@@ -3,8 +3,8 @@ package org.cryptimeleon.incentive.crypto;
 import org.cryptimeleon.craco.protocols.arguments.fiatshamir.FiatShamirProofSystem;
 import org.cryptimeleon.incentive.crypto.model.SpendRequest;
 import org.cryptimeleon.incentive.crypto.model.SpendResponse;
-import org.cryptimeleon.incentive.crypto.proof.spend.MetadataZkp;
 import org.cryptimeleon.incentive.crypto.proof.spend.SpendHelper;
+import org.cryptimeleon.incentive.crypto.proof.spend.zkp.MetadataZkp;
 import org.cryptimeleon.math.structures.cartesian.Vector;
 import org.cryptimeleon.math.structures.rings.zn.Zn;
 import org.junit.jupiter.api.Test;
